main/requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dnspython==2.5.0
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
future==0.18.3
hub-sdk==0.0.3
idna==3.6
iso8601==2.1.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
psutil==5.9.7
py-cpuinfo==9.0.0
pymongo==4.6.1
pyparsing==3.1.1
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
RPi.GPIO==0.7.1
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.1
urllib3==2.1.0
